% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstJulia.R
\name{firstJulia}
\alias{firstJulia}
\title{firstJulia}
\usage{
firstJulia()
}
\value{
It return the linked function of Julia
}
\description{
First step to check and start using Julia function for tracking tiny insect
}
\examples{

 #Just type the function
firstJulia()

}
