% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_Batch.R
\name{trackJR_Batch}
\alias{trackJR_Batch}
\title{trackJR_Batch}
\usage{
trackJR_Batch(pathDir)
}
\arguments{
\item{pathDir}{Path of the directory where videos saved.is a Character string; e.g. "C:myPC/videos". write the string without last bar "/"}
}
\value{
(list) It return the tracked directory in a list with all the videos with 3 column, the frame with x and y position. Be careful with X and Y position in the video, frames could be had (0,0) origin position at top-left of the plot.
}
\description{
Tracking a video of tiny insect to get frame, position X and position Y.In a batch mode.
}
\examples{
mypathDir<-"C:/Users/test" # write the string without last bar "/"
trackJR_Batch(mypathDir)

}
