% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_Batch.R
\name{trackJR_Batch_ggplot}
\alias{trackJR_Batch_ggplot}
\title{trackJR_Batch_ggplot}
\usage{
trackJR_Batch_ggplot(pathDir, dataTrackJR_list)
}
\arguments{
\item{pathDir}{Path of the directory where videos saved.is a Character string; e.g. "C:myPC/videos". write the string without last bar "/"}

\item{dataTrackJR_list}{The list with tracked videos.}
}
\value{
A list with the ggplot as an element of each video.
}
\description{
With this function you could add to the list the ggplot using `trackJR_ggplot`
}
\examples{
mypathDir<-"C:/Users/test" # write the string without last bar "/"
trackdata<-trackJR_Batch(mypathDir)
trackdata2<-trackJR_Batch_ggplot(mypathDir,trackdata)
trackdata2 #Now it has the ggplot as an element.
}
