% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_box.R
\name{trackJR_box}
\alias{trackJR_box}
\title{trackJR_box}
\usage{
trackJR_box(trackJRgg, var_name)
}
\arguments{
\item{trackJRgg}{A trackJR_ggplot.}

\item{var_name}{A character vector with the name of the box wanted (i.e. "box1")}
}
\value{
The function return the object named with the coordinates for the box or region of interest. A data.frame with one observation in four variables (xmin,xmax,ymin,ymax)
}
\description{
A function to create a coordinates from a box draw in a trackJR_ggplot. While brushing in the plot a box you can move and drag the box until presh the "finish" bottom
}
\examples{
trackJR_box(graf2,"box1")
#Use it in a ggplot generated with trackJR_ggplot
graf2 + geom_rect(data = box1,...)
}
