% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_clean.R
\name{trackJR_clean}
\alias{trackJR_clean}
\title{trackJR_clean}
\usage{
trackJR_clean(trackJRgg)
}
\arguments{
\item{trackJRgg}{The function need the trackJR_ggplot generated to select the point.}
}
\value{
The function return two object. A new data frame with points corrected at it previous frame for each point in problems. Also a data frame called "points_in_problem" is created.
}
\description{
Track cleaning function to solve points in problems. Points selected will have the last previous correct position tracked.
}
\examples{
#'mypathDirandFile<-"C:/Users/test/vid.mp4"
dataT<-trackJR(mypathDirandFile)
graf<-trackJR_ggplot(mypathDirandFile,dataT)

new_dataT<-trackJR_clean(graf)

}
