% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_pbox.R
\name{trackJR_pbox}
\alias{trackJR_pbox}
\title{trackJR_pbox}
\usage{
trackJR_pbox(data_trackJR, data_box)
}
\arguments{
\item{data_trackJR}{The data frame with column "x" and "y" generated by the trackJR function.}

\item{data_box}{The box coordinates generated by the [trackJR_box] to determine the point inside the box.}
}
\value{
a subset of the [data_trackJR] inside the box.
}
\description{
A function to select points inside a box.
}
\examples{
#' #'mypathDirandFile<-"C:/Users/test/vid.mp4"
dataT<-trackJR(mypathDirandFile)
graf<-trackJR_ggplot(mypathDirandFile,dataT)
trackJR_box(graf2,"box1")
estimuli<-trackJR_pbox(dataT,box1)
}
