% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_pref.R
\name{trackJR_pref}
\alias{trackJR_pref}
\title{trackJR_pref}
\usage{
trackJR_pref(graf, var_name)
}
\arguments{
\item{graf}{the ggplot plot generated with `trackJR_ggplot`}

\item{var_name}{the name added to the reference point.}
}
\value{
the reference points
}
\description{
trackJR_pref
}
\examples{
it is used internally in `trackJR_refpoint`
}
