% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_summ.R
\name{trackJR_summ}
\alias{trackJR_summ}
\title{trackJR_summ}
\usage{
trackJR_summ(data_trackJR, boxes = NULL, var_names = NULL)
}
\arguments{
\item{data_trackJR}{The data frame with columns "fr", "x" and "y" generated by the `trackJR` function.}

\item{boxes}{a Lit of The boxes names generated by the `trackJR_box` function.}

\item{var_names}{a List variable names to call the boxes.}
}
\value{
It return a data.frame with the count of point in each boxes.
}
\description{
Generate a simple summary for each `boxes` solving how many points are in the box?.
}
\examples{
mypathDirandFile<-"C:/Users/test/vid.mp4"
dataT<-trackJR(mypathDirandFile)
graf<-trackJR_ggplot(mypathDirandFile,dataT)
trackJR_box(graf,"box1")
trackJR_box(graf,"box2")
trackJR_box(graf,"box3")

summaryData<- trackJR_summ (dataT,list(box1,box2,box3),list("estimuli","control","blank"))

}
