% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackJR_ggplot.R
\name{trackJR_ggplot}
\alias{trackJR_ggplot}
\title{trackJR_ggplot}
\usage{
trackJR_ggplot(pathDirtoVid, dataTrackJR, frameIMG = 1)
}
\arguments{
\item{pathDirtoVid}{Character string for path to the video.}

\item{dataTrackJR}{The R object of your workspace to plot the track.}

\item{frameIMG}{The frame you want to plot.}
}
\value{
A ggplot object.
}
\description{
Generate a ggplot class object for the tracking a video. As a ggplot object you can add more layers.
}
\examples{
#'mypathDirandFile<-"C:/Users/test/vid.mp4"
dataT<-trackJR(mypathDirandFile)
graf<-trackJR_ggplot(mypathDirandFile,dataA)
}
